Separate frontend and backend logic, add Docker and docker-compose support
This commit took several steps to separate the frontend and backend logic, moving towards a microservices architecture. The backend now includes a separate Dockerfile and its dependencies while the frontend is moved to its folder with its unique Dockerfile and dependencies. The server is designed to return a list of movies and specific movies upon request. The frontend, on the other hand, makes requests to the backend and displays a list of movies and an individual movie upon selection.

The routing is modified to make direct requests to the backend with the help of environment variables. Docker-compose becomes part of our application to manage the two services and make it easy for anyone to spin up the whole application without worrying about dependencies and how to run the application.

Additionally, several files are renamed, and a new .env file is created in the frontend folder to hold environment variables. This change helps to decouple our application further and accommodate future scalability and maintainability.